Fix mobile media review feedback
