+++ mobile/lib/features/channels/channel_mutes/channel_mutes_manager.dart
@@ -8,7 +8,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../../shared/crypto/nip44.dart';
 import '../../../shared/relay/relay.dart';
-import '../read_state/read_state_time.dart';
+import '../../../shared/read_state/read_state_time.dart';
 import 'channel_mutes_storage.dart';
 
 class ChannelMutesCrypto {
