+++ mobile/lib/features/channels/message_actions.dart
@@ -25,9 +25,9 @@ import 'channel_management_provider.dart';
 import 'emoji_picker.dart';
 import 'reaction_row.dart';
 import 'recent_emoji_provider.dart';
-import 'read_state/message_read_state.dart';
-import 'read_state/read_state_format.dart';
-import 'read_state/read_state_provider.dart';
+import '../../shared/read_state/message_read_state.dart';
+import '../../shared/read_state/read_state_format.dart';
+import '../../shared/read_state/read_state_provider.dart';
 import 'thread_detail_page.dart';
 import 'thread_follows/thread_follows_provider.dart';
 import 'timeline_message.dart';
