+++ mobile/lib/shared/read_state/read_state_manager.dart
@@ -6,8 +6,8 @@ import 'package:flutter/foundation.dart';
 import 'package:nostr/nostr.dart' as nostr;
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../../../shared/crypto/nip44.dart';
-import '../../../shared/relay/relay.dart';
+import '../crypto/nip44.dart';
+import '../relay/relay.dart';
 import 'read_state_format.dart';
 import 'read_state_storage.dart';
 import 'read_state_time.dart';
