+++ mobile/test/features/channels/read_state/read_state_manager_test.dart
@@ -4,8 +4,8 @@ import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:nostr/nostr.dart' as nostr;
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:buzz/features/channels/read_state/read_state_format.dart';
-import 'package:buzz/features/channels/read_state/read_state_manager.dart';
+import 'package:buzz/shared/read_state/read_state_format.dart';
+import 'package:buzz/shared/read_state/read_state_manager.dart';
 import 'package:buzz/shared/relay/relay.dart';
 
 void main() {
