+++ mobile/test/features/channels/read_state/read_state_provider_test.dart
@@ -1,5 +1,5 @@
-import 'package:buzz/features/channels/read_state/read_state_format.dart';
-import 'package:buzz/features/channels/read_state/read_state_provider.dart';
+import 'package:buzz/shared/read_state/read_state_format.dart';
+import 'package:buzz/shared/read_state/read_state_provider.dart';
 import 'package:buzz/shared/community/community_provider.dart';
 import 'package:buzz/shared/relay/relay.dart';
 import 'package:buzz/shared/theme/theme_provider.dart';
